@@ -6,7 +6,7 @@
 
     <groupId>de.nehlen</groupId>
     <artifactId>Bingo</artifactId>
-    <version>1.3.0-RC2</version>
+    <version>1.4.3</version>
     <packaging>jar</packaging>
 
     <name>Bingo</name>
@@ -24,8 +24,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>9</source>
-                    <target>9</target>
+                    <source>16</source>
+                    <target>16</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -72,7 +72,7 @@
         <dependency>
             <groupId>io.papermc.paper</groupId>
             <artifactId>paper-api</artifactId>
-            <version>1.19.3-R0.1-SNAPSHOT</version>
+            <version>1.20.1-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
@@ -85,22 +85,25 @@
             <groupId>com.zaxxer</groupId>
             <artifactId>HikariCP</artifactId>
             <version>5.0.1</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>de.nehlen</groupId>
-            <artifactId>GameAPI</artifactId>
-            <version>1.4.0</version>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>20240303</version>
+        </dependency>
 
+        <dependency>
+            <groupId>de.nehlen</groupId>
+            <artifactId>Spookly-Core-API</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+            <scope>provided</scope>
+        </dependency>
         <dependency>
             <groupId>de.nehlen</groupId>
             <artifactId>SpooklyCloudNetUtils</artifactId>
-            <version>1.0</version>
-            <scope>system</scope>
-            <systemPath>P:/Spookly/Plugins/SpooklyCloudNetUtils/target/SpooklyCloudNetUtils-1.1.jar</systemPath>
+            <version>1.2</version>
+            <scope>provided</scope>
         </dependency>
-
     </dependencies>
 </project>
