@@ -4,11 +4,10 @@ import de.nehlen.bingo.Bingo;
 import de.nehlen.bingo.data.GameData;
 import de.nehlen.bingo.data.StringData;
 import de.nehlen.bingo.factory.UserFactory;
-import de.nehlen.gameapi.Gameapi;
-import de.nehlen.gameapi.PointsAPI.PointsAPI;
-import de.nehlen.gameapi.TeamAPI.Team;
+import de.nehlen.spookly.Spookly;
+import de.nehlen.spookly.player.SpooklyPlayer;
+import de.nehlen.spookly.team.Team;
 import lombok.Getter;
-import lombok.Setter;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Bukkit;
@@ -17,8 +16,6 @@ import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
-import java.util.stream.Collectors;
 
 public class PickList {
 
@@ -30,9 +27,11 @@ public class PickList {
     }
 
     public void completeMaterial(Player player, Team team, Material material) {
+        SpooklyPlayer spooklyPlayer = Spookly.getPlayer(player);
+
         if (!GameData.firstItemFound) {
-            Gameapi.getGameapi().getPointsAPI().updatePoints(player, PointsAPI.UpdateType.ADD, 100);
-            player.sendMessage(TextComponentHelper.addPointsComponent(100));
+            spooklyPlayer.addPoints(20);
+            player.sendMessage(TextComponentHelper.addPointsComponent(20));
             player.sendMessage(StringData.getPrefix()
                     .append(Component.text("Du hast den ").color(NamedTextColor.GRAY)
                             .append(Component.text("ersten Gegenstand").color(StringData.getHighlightColor()))
@@ -40,21 +39,18 @@ public class PickList {
             GameData.firstItemFound=true;
         }
 
-        Gameapi.getGameapi().getPointsAPI().updatePoints(player, PointsAPI.UpdateType.ADD, 100);
-        player.sendMessage(StringData.getPrefix()
-                .append(Component.text("Du hast ").color(NamedTextColor.GRAY))
-                .append(Component.text("100").color(NamedTextColor.RED))
-                .append(Component.text(" Punkte erhalten.").color(NamedTextColor.GRAY)));
+        spooklyPlayer.addPoints(20);
+        player.sendMessage(TextComponentHelper.addPointsComponent(20));
         Bingo.getBingo().getUserFactory().updateCraftedItems(player, UserFactory.UpdateType.ADD, 1);
         items.remove(material);
 
         Bukkit.broadcast(StringData.getPrefix()
-                .append(StringData.playerTeamPrefix(player))
-                .append(player.displayName())
-                .append(Component.text(" hat ").color(NamedTextColor.GRAY))
-                .append(Component.translatable(TranslatableHelper.getTranslationKey(material)).color(StringData.getHighlightColor()))
-                .append(Component.text(" gefunden.").color(NamedTextColor.GRAY))
-                .append(Component.text(" (" + getAmountCompleted() + "/" + GameData.getItemsAmount() + " Items)").color(NamedTextColor.GRAY)));
+                .append(Component.translatable("phase.ingame.itemFound",
+                        spooklyPlayer.nameTag(),
+                        Component.translatable(TranslatableHelper.getTranslationKey(material)).color(StringData.getHighlightColor()),
+                        Component.text(getAmountCompleted()),
+                        Component.text(GameData.getItemsAmount())
+                        ).color(NamedTextColor.GRAY)));
         if (isComplete()) {
             Bingo.getBingo().getIngameCountdown().endPhase();
             Bingo.getBingo().getEndingCoutdown().teamWin(team);
@@ -79,7 +75,7 @@ public class PickList {
         items.forEach(material -> {
             stringBuilder.append(material.toString() + ",");
         });
-        return stringBuilder.toString();
+        return "PickList[" + stringBuilder.toString() + "]";
     }
 
 }
