@@ -4,6 +4,9 @@ import de.nehlen.bingo.Bingo;
 import de.nehlen.bingo.data.GameData;
 import de.nehlen.bingo.data.GameState;
 import de.nehlen.bingo.data.StringData;
+import de.nehlen.spookly.Spookly;
+import de.nehlen.spookly.player.SpooklyPlayer;
+import de.nehlen.spookly.team.Team;
 import io.papermc.paper.event.player.AsyncChatEvent;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
@@ -24,27 +27,30 @@ public class AsyncPlayerChatListener implements Listener {
     public void onChat(AsyncChatEvent e) {
 
         Player p = e.getPlayer();
+        SpooklyPlayer spooklyPlayer = Spookly.getPlayer(p);
 
         e.setCancelled(true);
 
-        for(Player players: Bukkit.getOnlinePlayers()) {
-            if(GameState.state != GameState.END) {
-                if(!GameData.getIngame().contains(p)) {
-                    if(!GameData.getIngame().contains(players))	{
+        for (Player players : Bukkit.getOnlinePlayers()) {
+
+            if (GameState.state != GameState.END) {
+                if (!GameData.getIngame().contains(p)) {
+                    if (!GameData.getIngame().contains(players)) {
                         players.sendMessage(Component.text("✘ ").color(NamedTextColor.DARK_RED)
-                                .append(p.displayName())
+                                .append(spooklyPlayer.nameTag())
                                 .append(Component.text(": ").color(NamedTextColor.GRAY))
-                                .append(e.message()));
+                                .append(Spookly.getPlaceholderManager().replacePlaceholder(e.message(), p)));
                     }
                 } else {
-                    players.sendMessage(StringData.playerTeamPrefix(p).append(p.displayName())
+                    Team team = GameData.getTeamCache().get(players);
+                    players.sendMessage(spooklyPlayer.nameTag()
                             .append(Component.text(": ").color(NamedTextColor.GRAY))
-                            .append(e.message()));
+                            .append(Spookly.getPlaceholderManager().replacePlaceholder(e.message(), p)));
                 }
             } else {
-                players.sendMessage(StringData.playerTeamPrefix(p).append(p.displayName())
+                players.sendMessage(spooklyPlayer.nameTag()
                         .append(Component.text(": ").color(NamedTextColor.GRAY))
-                        .append(e.message()));
+                        .append(Spookly.getPlaceholderManager().replacePlaceholder(e.message(), p)));
             }
         }
     }
