@@ -4,8 +4,8 @@ import de.nehlen.bingo.Bingo;
 import de.nehlen.bingo.data.StringData;
 import de.nehlen.bingo.data.helper.TextComponentHelper;
 import de.nehlen.bingo.factory.UserFactory;
-import de.nehlen.gameapi.Gameapi;
-import de.nehlen.gameapi.PointsAPI.PointsAPI;
+import de.nehlen.spookly.Spookly;
+import de.nehlen.spookly.player.SpooklyPlayer;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.DyeColor;
@@ -31,15 +31,17 @@ public class EntityDeathListener implements Listener {
         Entity killer = event.getEntity().getKiller();
         if (killer instanceof Player) {
             Player player = (Player) killer;
+            SpooklyPlayer spooklyPlayer = Spookly.getPlayer(player);
+
             if (event.getEntityType().equals(EntityType.SHEEP)) {
                 Sheep sheep = (Sheep) entity;
                 if (sheep.getColor().equals(DyeColor.BROWN)) {
                     player.sendMessage(StringData.getPrefix()
-                            .append(Component.text("Du hast ein ").color(NamedTextColor.GRAY)
-                                    .append(Component.text("Braunes-Schaf").color(StringData.getHighlightColor()))
-                                    .append(Component.text(" getötet.").color(NamedTextColor.GRAY))));
-                    player.sendMessage(TextComponentHelper.addPointsComponent(100));
-                    Gameapi.getGameapi().getPointsAPI().updatePoints(player, PointsAPI.UpdateType.ADD, 100);
+                            .append(Component.translatable("stats.brownSheep.add",
+                                    Component.translatable("stats.brownSheep.sheep").color(StringData.getHighlightColor())
+                            ).color(NamedTextColor.GRAY)));
+                    player.sendMessage(TextComponentHelper.addPointsComponent(20));
+                    spooklyPlayer.addPoints(20);
                     this.bingo.getUserFactory().updateBrownSheeps(player, UserFactory.UpdateType.ADD, 1);
                 }
             }
