@@ -4,8 +4,8 @@ import de.nehlen.bingo.Bingo;
 import de.nehlen.bingo.data.GameData;
 import de.nehlen.bingo.data.GameState;
 import de.nehlen.bingo.data.StringData;
-import de.nehlen.gameapi.Gameapi;
-import de.nehlen.gameapi.TeamAPI.Team;
+import de.nehlen.spookly.Spookly;
+import de.nehlen.spookly.team.Team;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Bukkit;
@@ -27,15 +27,21 @@ public class PlayerQuitListener implements Listener {
         try {
             Player player = event.getPlayer();
             Component quitMessage = StringData.getPrefix()
-                    .append(event.getPlayer().name().color(StringData.getHighlightColor()))
-                    .append(Component.text(" hat das Spiel verlassen.").color(NamedTextColor.GRAY));
+                    .append(Component.translatable("general.quit.message",
+                            event.getPlayer().displayName().color(StringData.getHighlightColor())
+                    ).color(NamedTextColor.GRAY));
 
             if (GameState.state == GameState.LOBBY) {
                 event.quitMessage(quitMessage);
                 if (Bukkit.getOnlinePlayers().size() <= GameData.getMinPlayerToStartGame()) {
                     Bukkit.broadcast(StringData.getPrefix()
-                            .append(Component.text("Der Countdown wurde gestopt weil zu wenig Spieler drin sind.").color(NamedTextColor.GRAY)));
-                    Bingo.getBingo().getLobbyCountdown().endPhase();
+                            .append(Component.translatable("phase.lobby.countdownStop").color(NamedTextColor.GRAY)));
+                    Bingo.getBingo().getLobbyPhase().counter(GameData.getStartTime());
+                    Bingo.getBingo().getLobbyPhase().endPhase();
+                }
+
+                if (GameData.getTeamCache().containsKey(player)) {
+                    GameData.getTeamCache().get(player).removePlayer(player);
                 }
             } else if (GameState.state == GameState.INGAME && GameData.getIngame().contains(player)) {
                 event.quitMessage(quitMessage);
@@ -43,19 +49,15 @@ public class PlayerQuitListener implements Listener {
                 Team team = GameData.getTeamCache().get(player);
                 team.removePlayer(player);
                 GameData.getTeamCache().remove(player);
-                if(team.registeredPlayers().isEmpty()) {
-                    Gameapi.getGameapi().getTeamAPI().removeTeam(team);
+                if (team.registeredPlayers().isEmpty()) {
+                    Spookly.getTeamManager().removeTeam(team);
                 }
 
-                if (Gameapi.getGameapi().getTeamAPI().registeredTeams().size() == 1) {
-                    Bingo.getBingo().getEndingCoutdown().teamWin(GameData.getTeamCache().get(GameData.getIngame().get(0)));
+                if (Spookly.getTeamManager().registeredTeams().size() == 1) {
+                    Bingo.getBingo().getEndingCoutdown().teamWin(Spookly.getTeamManager().registeredTeams().get(0));
                 }
             } else {
                 event.quitMessage(Component.empty());
-            }
-
-            if(GameData.getTeamCache().containsKey(player)) {
-                GameData.getTeamCache().get(player).removePlayer(player);
             }
         } catch (Exception e) {
 //            Sentry.captureException(e);
