@@ -1,32 +1,39 @@
 package de.nehlen.bingo.phases;
 
 import de.nehlen.bingo.Bingo;
+import de.nehlen.bingo.bossbar.BossComponentHelper;
+import de.nehlen.bingo.commands.hudCommand;
 import de.nehlen.bingo.data.GameData;
 import de.nehlen.bingo.data.GameState;
 import de.nehlen.bingo.data.StringData;
 import de.nehlen.bingo.data.helper.PickList;
 import de.nehlen.bingo.data.helper.TextComponentHelper;
-import de.nehlen.bingo.util.Items;
+import de.nehlen.bingo.util.AbstractGamePhase;
 import de.nehlen.bingo.util.UtilFunctions;
-import de.nehlen.gameapi.Gameapi;
-import de.nehlen.gameapi.PhaseApi.AbstractGamePhase;
-import de.nehlen.gameapi.TeamAPI.Team;
+import de.nehlen.spookly.Spookly;
+import de.nehlen.spookly.team.Team;
+import net.kyori.adventure.bossbar.BossBar;
+import net.kyori.adventure.key.Key;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
-import net.kyori.adventure.text.format.TextDecoration;
 import org.bukkit.Bukkit;
-import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class IngameCountdown extends AbstractGamePhase {
     private final Bingo bingo;
+    private BossBar bossBar;
+    private Map<Player, BossBar> bossBars = new HashMap<>();
 
     public IngameCountdown(Bingo bingo) {
         super(GameData.getMaxGameTime());
         this.bingo = bingo;
+
+        bossBar = BossBar.bossBar(Component.empty(), 0, BossBar.Color.WHITE, BossBar.Overlay.NOTCHED_20);
     }
 
     public void startPhase() {
@@ -35,13 +42,34 @@ public class IngameCountdown extends AbstractGamePhase {
         if (GameData.getActiveNether())
             bingo.getWorldManager().setWorldSettingsForGameWorlds(Objects.requireNonNull(Bukkit.getWorld("world_nether")));
 
-        Bukkit.getOnlinePlayers().forEach(Bingo.getBingo().getScoreboardManager()::setUserScoreboard);
+        Bukkit.getOnlinePlayers().forEach(player -> {
+            Bingo.getBingo().getScoreboardManager().setUserScoreboard(player);
+            if (hudCommand.getBetaPlayer().contains(player))
+                player.showBossBar(bossBar);
+        });
+
 
         scheduler = Bukkit.getScheduler().scheduleSyncRepeatingTask(Bingo.getBingo(), () -> {
+//            bossBar.name(
+//                    Component.empty()
+//                            .append(BossComponentHelper.container(BossComponentHelper.BossBackgroundSize.SIZE_64,
+//                                    Component.empty()
+//                                            .append(Component.text('\uE100').font(Key.key("hud")).color(BossComponentHelper.getNoShadowColor()))
+//                                            .append(Component.text('\uEFE2').font(Key.key("hud")))
+//                                            .append(Component.text(Bingo.getBingo().getScoreboardManager().getTeam(player)))))
+//                            .append(Component.text('\uEFE6').font(Key.key("hud"))) //64 spacer
+//                            .append(BossComponentHelper.container(BossComponentHelper.BossBackgroundSize.SIZE_64,
+//                                    Component.empty() //Time
+//                                            .append(Component.text('\uE100').font(Key.key("hud")).color(BossComponentHelper.getNoShadowColor()))
+//                                            .append(Component.text('\uEFE2').font(Key.key("hud")))
+//                                            .append(Component.text(UtilFunctions.formatTime(getCounter())))))
+//            );
+
+
             Bukkit.getScheduler().runTaskAsynchronously(bingo, () -> {
                 Bukkit.getOnlinePlayers().forEach(player -> {
-
-                    player.sendActionBar(Bingo.getBingo().getScoreboardManager().getTeam(player)
+                    player.sendActionBar(Component.empty()
+                            .append(Bingo.getBingo().getScoreboardManager().getTeam(player))
                             .append(TextComponentHelper.seperator())
                             .append(Component.text(UtilFunctions.formatTime(getCounter())).color(NamedTextColor.WHITE)));
                 });
@@ -50,7 +78,7 @@ public class IngameCountdown extends AbstractGamePhase {
 
             if (counter == 0) {
                 Map<Team, Integer> teamItems = new HashMap<>();
-                Gameapi.getGameapi().getTeamAPI().registeredTeams().forEach(team -> {
+                Spookly.getTeamManager().registeredTeams().forEach(team -> {
                     teamItems.put(team, 9 - ((PickList) team.memory().get("picklist")).getAmountCompleted());
 
                 });
@@ -62,16 +90,7 @@ public class IngameCountdown extends AbstractGamePhase {
                 bingo.getEndingCoutdown().teamWin(winner);
                 Bukkit.broadcast(StringData.getPrefix()
                         .append(Component.text("Die Spielzeit ist abgelaufen, das Team mit den meißten Items hat gewonnen.").color(NamedTextColor.GRAY)));
-
-//                for (Player player : Bukkit.getOnlinePlayers()) {
-//                    player.teleport(GameData.getLobbyLocation());
-//                    player.getInventory().clear();
-//                    player.getInventory().setItem(8, Items.createItem(Material.HEART_OF_THE_SEA, Component.text("Zurück zur Lobby").color(NamedTextColor.GRAY), 1));
-//                }
-//                GameState.state = GameState.END;
-//                Bingo.getBingo().getEndingCoutdown().startPhase();
             }
-
             counter--;
         }, 20L, 20L);
     }
